Tweaked behavior of the Prompt to feel more realistic
The input text field takes focus when the page first loads.
The input text field's contents are wiped after a "command" is entered.
The input text field does not allow for autocorrecting, autocapitalization, or anything of the sort.
A dollar sign prompt appears before each command that you type.
The input text field lines up well with the dollar sign prompt.
The input text field takes up all available width of the column that it's in.
Removed specified background colors from both columns on the webpage.